rename event dashboard
