+++ frontend/src/components.d.ts
@@ -12,11 +12,9 @@ declare module "@vue/runtime-core" {
     EditEvent: typeof import("./components/EditEvent.vue")["default"];
     EventCreatePage: typeof import("./components/pages/EventCreatePage.vue")["default"];
     EventDashboard: typeof import("./components/pages/EventDashboard.vue")["default"];
-    EventDashboardOngoing: typeof import("./components/pages/EventDashboardOngoing.vue")["default"];
     EventEditPage: typeof import("./components/pages/EventEditPage.vue")["default"];
     EventInfo: typeof import("./components/EventInfo.vue")["default"];
     EventListItem: typeof import("./components/EventListItem.vue")["default"];
-    EventOrganizerView: typeof import("./components/EventOrganizerView.vue")["default"];
     EventPage: typeof import("./components/pages/EventPage.vue")["default"];
     EventSearchPage: typeof import("./components/pages/EventSearchPage.vue")["default"];
     HomePage: typeof import("./components/pages/HomePage.vue")["default"];
