+++ frontend/src/components/pages/EventDashboard.vue
